Implement 'ParquetFile._write_common_metadata()'.
