+++ fastparquet/__init__.py
@@ -1,9 +1,8 @@
 """parquet - read parquet files."""
 __version__ = "0.7.1"
 
-from .thrift_structures import parquet_thrift
+from .thrift_structures import parquet_thrift, ParquetException
 from .core import read_thrift
 from .writer import write
 from . import core, schema, converted_types, api
 from .api import ParquetFile
-from .util import ParquetException
