+++ fastparquet/util.py
@@ -1,12 +1,13 @@
-import copy
+from copy import copy
 import io
-import struct
+import json
 import numpy as np
 import os
 import os.path
 import operator
 import pandas as pd
 import re
+import struct
 import numbers
 from collections import defaultdict
 from distutils.version import LooseVersion
@@ -15,14 +16,11 @@ import pandas
 
 from pandas.api.types import is_categorical_dtype
 
+from .thrift_structures import write_thrift
+
 PANDAS_VERSION = LooseVersion(pandas.__version__)
+MARKER = b'PAR1'
 created_by = "fastparquet-python version 1.0.0 (build 111)"
-
-
-class ParquetException(Exception):
-    """Generic Exception related to unexpected data format when
-     reading parquet file."""
-    pass
 
 
 def default_mkdirs(f):
@@ -112,6 +110,52 @@ def check_column_names(columns, *args):
                                  "" % (missing, arg, columns))
 
 
+def write_common_metadata(fn, fmd, open_with=default_open,
+                          no_row_groups=True):
+    """
+    For hive-style parquet, write schema in special shared file
+
+    Parameters
+    ----------
+    fn: str
+        Filename to write to
+    fmd: thrift FileMetaData
+        Information to write
+    open_with: func
+        To use to create writable file as f(path, mode)
+    no_row_groups: bool (True)
+        Strip out row groups from metadata before writing - used for "common
+        metadata" files, containing only the schema.
+    """
+    consolidate_categories(fmd)
+    with open_with(fn, 'wb') as f:
+        f.write(MARKER)
+        if no_row_groups:
+            fmd = copy(fmd)
+            fmd.row_groups = []
+        foot_size = write_thrift(f, fmd)
+        f.write(struct.pack(b"<i", foot_size))
+        f.write(MARKER)
+
+
+def consolidate_categories(fmd):
+    key_value = [k for k in fmd.key_value_metadata
+                 if k.key == 'pandas'][0]
+    meta = json.loads(key_value.value)
+    cats = [c for c in meta['columns']
+            if 'num_categories' in (c['metadata'] or [])]
+    for cat in cats:
+        for rg in fmd.row_groups:
+            for col in rg.columns:
+                if ".".join(col.meta_data.path_in_schema) == cat['name']:
+                    ncats = [k.value for k in (col.meta_data.key_value_metadata or [])
+                             if k.key == 'num_categories']
+                    if ncats and int(ncats[0]) > cat['metadata'][
+                            'num_categories']:
+                        cat['metadata']['num_categories'] = int(ncats[0])
+    key_value.value = json.dumps(meta, sort_keys=True)
+
+
 def metadata_from_many(file_list, verify_schema=False, open_with=default_open,
                        root=False, fs=None):
     """
@@ -176,22 +220,22 @@ def metadata_from_many(file_list, verify_schema=False, open_with=default_open,
                 if pf._schema != pfs[0]._schema:
                     raise ValueError('Incompatible schemas')
 
-        fmd = copy.copy(pfs[0].fmd)  # we inherit "created by" field
+        fmd = copy(pfs[0].fmd)  # we inherit "created by" field
         fmd.row_groups = []
 
         for pf, fn in zip(pfs, file_list):
             if pf.file_scheme not in ['simple', 'empty']:
                 for rg in pf.row_groups:
-                    rg = copy.copy(rg)
-                    rg.columns = [copy.copy(c) for c in rg.columns]
+                    rg = copy(rg)
+                    rg.columns = [copy(c) for c in rg.columns]
                     for chunk in rg.columns:
                         chunk.file_path = '/'.join([fn, chunk.file_path])
                     fmd.row_groups.append(rg)
 
             else:
                 for rg in pf.row_groups:
-                    rg = copy.copy(rg)
-                    rg.columns = [copy.copy(c) for c in rg.columns]
+                    rg = copy(rg)
+                    rg.columns = [copy(c) for c in rg.columns]
                     for chunk in rg.columns:
                         chunk.file_path = fn
                     fmd.row_groups.append(rg)
